ocr_pipeline/requirements.txt
transformers
Pillow
pandas
openpyxl
